requirements.txt
python-ffmpeg
pyautogui
keyboard
py-cpuinfo
GPUtil
mss
numpy